unify virtual guest defaults used in API and WebUI
memory to 512MB
storage to 3GB

+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/system/SystemHandler.java
@@ -2134,8 +2134,8 @@ public class SystemHandler extends BaseHandler {
 
     /**
      *
-     * Provision a guest on the server specified.  Defaults to: memory=256MB, vcpu=1,
-     * storage=2GB.
+     * Provision a guest on the server specified.  Defaults to: memory=512, vcpu=1,
+     * storage=3GB.
      *
      * @param sessionKey of user making call
      * @param sid of server to provision guest on
@@ -2144,7 +2144,7 @@ public class SystemHandler extends BaseHandler {
      * @return Returns 1 if successful, exception otherwise
      *
      * @xmlrpc.doc Provision a guest on the host specified.  Defaults to:
-     * memory=256MB, vcpu=1, storage=2GB.
+     * memory=512MB, vcpu=1, storage=3GB.
      * @xmlrpc.param #param("string", "sessionKey")
      * @xmlrpc.param #param("int", "serverId") - ID of host to provision guest on.
      * @xmlrpc.param #param("string", "guestName")
@@ -2154,7 +2154,7 @@ public class SystemHandler extends BaseHandler {
     public int provisionVirtualGuest(String sessionKey, Integer sid, String guestName,
             String profileName) {
         return provisionVirtualGuest(sessionKey, sid, guestName, profileName,
-                new Integer(256), new Integer(1), new Integer(2));
+                new Integer(512), new Integer(1), new Integer(3));
     }
 
     /**
